refactor files - renderer another try
